src/requirements.txt
pipecat-ai[websocket,deepgram,groq,openai,silero]
fastapi
uvicorn[standard]
google-cloud-storage
python-dotenv
